MADE FUNCTIONS INTO HEADERS!! Added house, some new commands, added some stuff to README.md
